MOre fixes
